restructure taskeditor
